replace signal by a call to create_post_notification method
