+++ lacommunaute/forum_conversation/models.py
@@ -9,9 +9,9 @@ from machina.apps.forum_conversation.abstract_models import AbstractPost, Abstra
 from machina.models.abstract_models import DatedModel
 from taggit.managers import TaggableManager
 
-from lacommunaute.forum_conversation.signals import post_create
 from lacommunaute.forum_member.shortcuts import get_forum_member_display_name
 from lacommunaute.forum_upvote.models import UpVote
+from lacommunaute.notification.utils import create_post_notifications
 from lacommunaute.users.enums import EmailLastSeenKind
 from lacommunaute.users.models import EmailLastSeen, User
 
@@ -155,10 +155,11 @@ class Post(AbstractPost):
         created = not self.pk
         super().save(*args, **kwargs)
         if created:
-            post_create.send(sender=self.__class__, instance=self)
+            create_post_notifications(post=self)
             EmailLastSeen.objects.seen(
                 email=self.poster.email if self.poster else self.username, kind=EmailLastSeenKind.POST
             )
+
 
 
 class CertifiedPost(DatedModel):
