--- lacommunaute/forum_conversation/signals.py
@@ -1,4 +1,0 @@
-from django.db.models.signals import ModelSignal
-
-
-post_create = ModelSignal(use_caching=True)
