+++ lacommunaute/notification/utils.py
@@ -2,8 +2,8 @@ from datetime import datetime
 
 from django.utils.timezone import now, timedelta
 
-from lacommunaute.notification.enums import EmailSentTrackKind
-from lacommunaute.notification.models import EmailSentTrack
+from lacommunaute.notification.enums import EmailSentTrackKind, delay_of_notifications
+from lacommunaute.notification.models import EmailSentTrack, Notification
 from lacommunaute.users.models import User
 
 
@@ -29,3 +29,27 @@ def get_serialized_messages(notifications):
         }
         for n in notifications
     ]
+
+
+def create_post_notifications(post):
+    """
+    When a Post is created, we schedule notifications to users active in the thread
+    """
+    # NOTE: notifications solution depends on the fact that approval is completed at the time of creation
+    if not post.approved:
+        return
+
+    if post.is_topic_head:
+        kind = EmailSentTrackKind.PENDING_TOPIC
+    elif post.is_first_reply:
+        kind = EmailSentTrackKind.FIRST_REPLY
+    else:
+        kind = EmailSentTrackKind.FOLLOWING_REPLIES
+
+    delay = delay_of_notifications[kind]
+
+    notifications = [
+        Notification(recipient=email_address, post=post, kind=kind, delay=delay)
+        for email_address in post.topic.mails_to_notify()
+    ]
+    Notification.objects.bulk_create(notifications)
